requirements.txt
# AWS Lambda
boto3==1.34.14
aws-lambda-powertools==2.31.0
# Core dependencies
sentence-transformers>=2.2.2
umap-learn>=0.5.4
numpy>=1.21.0
scikit-learn>=1.0.2
matplotlib>=3.5.0
scipy>=1.7.0
pandas>=1.3.0

# Dependencies of sentence-transformers
torch>=1.11.0
transformers>=4.20.0
tqdm>=4.62.0
huggingface-hub>=0.4.0

# Dependencies of umap-learn
numba>=0.51.0
pynndescent>=0.5.4

# Optional but recommended for better performance
# Uncomment if needed:
# accelerate>=0.20.0  # For faster model inference
# faiss-cpu>=1.7.4   # For efficient similarity search with large datasets